Function build_drake_graph

Use drake_config() instead.

build_drake_graph(
  plan,
  targets = plan$target,
  envir = parent.frame(),
  verbose = 1L,
  jobs = 1,
  console_log_file = NULL,
  trigger = drake::trigger(),
  cache = NULL
)

Arguments

plan
Workflow plan data frame. A workflow plan data frame is a data frame with a target column and a command column. (See the details in the drake_plan() help file for descriptions of the optional columns.) Targets are the objects that drake generates, and commands are the pieces of R code that produce them. You can create and track custom files along the way (see file_in(), file_out(), and knitr_in()). Use the function drake_plan() to generate workflow plan data frames.

targets
Character vector, names of targets to build. Dependencies are built too. You may supply static and/or whole dynamic targets, but no sub-targets.

envir
Environment to use. Defaults to the current workspace, so you should not need to worry about this most of the time. A deep copy of envir is made, so you don't need to worry about your workspace being modified by make. The deep copy inherits from the global environment. Wherever necessary, objects and functions are imported from envir and the global environment and then reproducibly tracked as dependencies.

verbose
Integer, control printing to the console/terminal.

0: print nothing.

1: print target-by-target messages as make() progresses.

2: show a progress bar to track how many targets are done so far.

jobs
Maximum number of parallel workers for processing the targets. You can experiment with predict_runtime() to help decide on an appropriate number of jobs. For details, visit https://books.ropensci.org/drake/time.html.

console_log_file
Deprecated in favor of log_make.

trigger
Name of the trigger to apply to all targets. Ignored if plan has a trigger column. See trigger() for details.

cache
drake cache as created by new_cache(). See also drake_cache().

Value

An igraph object.

Details

Deprecated on 2018-11-02.