List the time it took to build each target/import.

Listed times do not include the amount of time spent loading and saving objects! See the type argument for different versions of the build time. (You can choose whether to take storage time into account.)

build_times(..., path = getwd(), search = TRUE, digits = 3,
  cache = get_cache(path = path, search = search, verbose = verbose),
  targets_only = FALSE, verbose = TRUE, jobs = 1, type = c("build",
  "command"))

Arguments

...
targets to load from the cache: as names (symbols), character strings, or dplyr-style tidyselect commands such as starts_with().

path
Root directory of the drake project, or if search is TRUE, either the project root or a subdirectory of the project.

search
logical. If TRUE, search parent directories to find the nearest drake cache. Otherwise, look in the current working directory only.

digits
How many digits to round the times to.

cache
optional drake cache. If supplied, the path and search arguments are ignored.

targets_only
logical, whether to only return the build times of the targets (exclude the imports).

verbose
whether to print console messages

jobs
number of parallel jobs/workers for light parallelism.

type
Type of time you want: either "build" for the full build time including the time it took to store the target, or "command" for the time it took just to run the command.

Value

A data frame of times, each from system.time().

See also

built()

Examples

# NOT RUN {
test_with_dir("Quarantine side effects.", {
# Show the build times for the basic example.
load_basic_example() # Get the code with drake_example("basic").
make(my_plan) # Build all the targets.
build_times() # Show how long it took to build each target.
build_times(starts_with("coef")) # `dplyr`-style `tidyselect`
})
# }