List targets in the cache.

Tip: read/load a cached item with readd() or loadd().

cached(..., list = character(0), no_imported_objects = FALSE,
  path = NULL, search = NULL, cache = drake_cache(path = path,
  verbose = verbose), verbose = 1L, namespace = NULL, jobs = 1,
  targets_only = TRUE)

Arguments

...
Deprecated. Do not use. Objects to load from the cache, as names (unquoted) or character strings (quoted). Similar to ... in remove().

list
Deprecated. Do not use. Character vector naming objects to be loaded from the cache. Similar to the list argument of remove().

no_imported_objects
Logical, deprecated. Use targets_only instead.

path
Path to a drake cache (usually a hidden .drake/ folder) or NULL.

search
Deprecated.

cache
drake cache. See new_cache(). If supplied, path is ignored.

verbose
Integer, control printing to the console/terminal.

0: print nothing.

1: print targets, retries, and failures.

2: also show a spinner when preprocessing tasks are underway.

namespace
Character scalar, name of the storr namespace to use for listing objects.

jobs
Number of jobs/workers for parallel processing.

targets_only
Logical. If TRUE just list the targets. If FALSE, list files and imported objects too.

Value

Either a named logical indicating whether the given targets or cached or a character vector listing all cached items, depending on whether any targets are specified.

See also

Examples

# NOT RUN {
isolate_example("Quarantine side effects.", {
if (suppressWarnings(require("knitr"))) {
if (requireNamespace("lubridate")) {
load_mtcars_example() # Load drake's canonical example.
make(my_plan) # Run the project, build all the targets.
cached()
cached(targets_only = FALSE)
}
}
})
# }