Return the default value of the args argument to make().

For make(..., parallelism = "Makefile"), this function configures the default arguments to system2(). It is an internal function, and most users do not need to worry about it.

default_Makefile_args(jobs, verbose)

Arguments

jobs
number of jobs

verbose
logical or numeric, control printing to the console. Use pkgconfig to set the default value of verbose for your R session: for example, pkgconfig::set_config("drake::verbose" = 2).

0 or FALSE:

print nothing.

1 or TRUE:

print only targets to build.

2:

in addition, print checks and cache info.

3:

in addition, print any potentially missing items.

4:

in addition, print imports. Full verbosity.

Value

args for system2(command, args)

Examples

default_Makefile_args(jobs = 2, verbose = FALSE)

#> [1] "--jobs=2" "--silent"

default_Makefile_args(jobs = 4, verbose = TRUE)

#> [1] "--jobs=4"