Download and save the code and data files of an example drake-powered project.

The drake_example() function downloads a folder from https://github.com/wlandau/drake-examples. By default, it creates a new folder with the example name in your current working directory. After the files are written, have a look at the enclosed README file. Other instructions are available in the files at https://github.com/wlandau/drake-examples.

drake_example(example = "main", to = getwd(), destination = NULL,
  overwrite = FALSE, quiet = TRUE)

Arguments

example
Name of the example. The possible values are the names of the folders at https://github.com/wlandau/drake-examples.

to
Character scalar, the folder containing the code files for the example. passed to the exdir argument of utils::unzip().

destination
Deprecated; use to instead.

overwrite
Logical, whether to overwrite an existing folder with the same name as the drake example.

quiet
Logical, passed to downloader::download() and thus utils::download.file(). Whether to download quietly or print progress.

Value

NULL

See also

Examples

# NOT RUN {
isolate_example("Quarantine side effects.", {
if (requireNamespace("downloader")) {
drake_examples() # List all the drake examples.
# Sets up the same example from load_mtcars_example()
drake_example("mtcars")
# Sets up the SLURM example.
drake_example("slurm")
}
})
# }