Task passed to individual futures in the "future" backend

For internal use only. Only exported to make available to futures.

drake_future_task(target, meta, config, protect)

Arguments

target
name of the target

meta
list of metadata

config
drake_config() list

protect
Names of targets that still need their dependencies available in config$envir.

Value

Either the target value or a list of build results.