Create replicates of targets.

expand_plan() is no longer recommended. Consider using transformations instead. Visit https://ropenscilabs.github.io/drake-manual/plans.html#large-plans for the details.

expand_plan(plan, values = NULL, rename = TRUE, sep = "_",
  sanitize = TRUE)

Arguments

plan
Workflow plan data frame.

values
Values to expand over. These will be appended to the names of the new targets.

rename
Logical, whether to rename the targets based on the values. See the examples for a demo.

sep
Character scalar, delimiter between the original target names and the values to append to create the new target names. Only relevant when rename is TRUE.

sanitize
Logical, whether to sanitize the plan.

Value

An expanded workflow plan data frame (with replicated targets).

Details

Duplicates the rows of a workflow plan data frame. Prefixes are appended to the new target names so targets still have unique names.

See also

Examples

# Create the part of the workflow plan for the datasets.
datasets <- drake_plan(
  small = simulate(5),
  large = simulate(50)
)
# Create replicates. If you want repeat targets,
# this is convenient.
expand_plan(datasets, values = c("rep1", "rep2", "rep3"))

#> # A tibble: 6 x 2
#>   target     command     
#>   <chr>      <S3: langs> 
#> 1 small_rep1 simulate(5) 
#> 2 small_rep2 simulate(5) 
#> 3 small_rep3 simulate(5) 
#> 4 large_rep1 simulate(50)
#> 5 large_rep2 simulate(50)
#> 6 large_rep3 simulate(50)

# Choose whether to rename the targets based on the values.
expand_plan(datasets, values = 1:3, rename = TRUE)

#> # A tibble: 6 x 2
#>   target  command     
#>   <chr>   <S3: langs> 
#> 1 small_1 simulate(5) 
#> 2 small_2 simulate(5) 
#> 3 small_3 simulate(5) 
#> 4 large_1 simulate(50)
#> 5 large_2 simulate(50)
#> 6 large_3 simulate(50)

expand_plan(datasets, values = 1:3, rename = FALSE)

#> # A tibble: 6 x 2
#>   target command     
#>   <chr>  <S3: langs> 
#> 1 small  simulate(5) 
#> 2 small  simulate(5) 
#> 3 small  simulate(5) 
#> 4 large  simulate(50)
#> 5 large  simulate(50)
#> 6 large  simulate(50)