Get a target's plan info from inside a plan's command.

Call this function inside the commands in your plan to get an entry from any column in the plan. Changes to custom columns referred to this way (for example, a in drake_plan(x = target(from_plan("a"), a = 123))) do not invalidate targets, so be careful. Only use from_plan() to reference data that does not actually affect the output value of the target. For example, you might use from_plan() to set the number of parallel workers within a target: drake_plan(x = target(mclapply(..., from_plan("cores"), cores = 4)). Now, if you change the cores column of the plan, the parallelism will change, but the target x will stay up to date.

from_plan(column)

Arguments

column
Character, name of a column in your drake plan.

Value

plan[target, column], where plan is your workflow plan data frame, target is the target being built, and column is the name of the column of the plan you provide.

See also

Examples

plan <- drake_plan(my_target = target(from_plan("a"), a = "a_value"))
plan

#> # A tibble: 1 x 3
#>   target    command            a      
#>   <chr>     <chr>              <chr>  
#> 1 my_target "from_plan(\"a\")" a_value

cache <- storr::storr_environment()
make(plan, cache = cache, session_info = FALSE)

#> target my_target

readd(my_target, cache = cache)

#> [1] "a_value"

# Why do we care?
# Because we can customize parallel computing within targets this way.
# Notice how we treat `mc.cores` for `mclapply()`
plan <- drake_plan(
  a = target(
    mclapply(1:8, my_function, mc.cores = from_plan("cores")),
    cores = 4
  ),
  b = target(
    mclapply(1:4, my_function, mc.cores = from_plan("cores")),
    cores = 2
  )
)

plan

#> # A tibble: 2 x 3
#>   target command                                                       cores
#>   <chr>  <chr>                                                         <dbl>
#> 1 a      "mclapply(1:8, my_function, mc.cores = from_plan(\"cores\"))"     4
#> 2 b      "mclapply(1:4, my_function, mc.cores = from_plan(\"cores\"))"     2

# make() # nolint

# Now, if you change the `cores` column of the plan, the parallelism will
# change, but the targets will stay up to date.
plan$cores <- c(1, 1)
plan

#> # A tibble: 2 x 3
#>   target command                                                       cores
#>   <chr>  <chr>                                                         <dbl>
#> 1 a      "mclapply(1:8, my_function, mc.cores = from_plan(\"cores\"))"     1
#> 2 b      "mclapply(1:4, my_function, mc.cores = from_plan(\"cores\"))"     1

# make(plan) # nolint