Deprecated. List the targets that either (1) are currently being built during a make(), or (2) were being built if the last make() quit unexpectedly.

Deprecated on 2019-01-13.

in_progress(path = getwd(), search = TRUE,
  cache = drake::get_cache(path = path, search = search, verbose =
  verbose), verbose = 1L)

Arguments

path
Root directory of the drake project, or if search is TRUE, either the project root or a subdirectory of the project. Ignored if a cache is supplied.

search
Logical. If TRUE, search parent directories to find the nearest drake cache. Otherwise, look in the current working directory only. Ignored if a cache is supplied.

cache
drake cache. See new_cache(). If supplied, path and search are ignored.

verbose
Integer, control printing to the console/terminal.

0: print nothing.

1: print targets, retries, and failures.

2: also show a spinner when preprocessing tasks are underway.

Value

A character vector of target names.

Details

Similar to progress().

See also