Map a function to a grid of settings to create a plan.

Given a function name and a data frame of named arguments, map_plan() creates a workflow plan data frame where each target is the result of the function applied to the arguments in a row. The idea is similar to functions map_df() and pmap()

map_plan(args, fun, id = "id", character_only = FALSE)

Arguments

args
a data frame of function arguments. Here, the column names should be the names of the arguments of fun, and each row of args corresponds to a call to fun.

fun
name of a function to apply the arguments row-by-row. Supply a symbol if character_only is FALSE and a character scalar otherwise.

id
name of an optional column in args giving the names of the targets. If not supplied, target names will be generated automatically. id should be a symbol if character_only is FALSE and a character scalar otherwise.

character_only
logical, whether to interpret the fun and id arguments as character scalars or symbols.

Value

A workflow plan data frame

See also

Examples

# For a practical example of map_plan() in action,
# see the econometrics exmaple at
# https://ropenscilabs.github.io/drake-manual/gsp.html.
# But for now, let's say we want find out which variables
# best explain fuel efficiency (in miles per gallon) in the mtcars dataset.
# We write a function that takes the names of two covariates
# and fits them to the response variable (mpg).
my_model_fit <- function(x1, x2){
  lm(as.formula(paste("mpg ~", x1, "+", x2)), data = mtcars)
}
# For covariates, we consider all the variables in mtcars except mpg.
covariates <- setdiff(colnames(mtcars), "mpg")
# And we consider all possible combinations of 2 covariates.
args <- as.data.frame(t(combn(covariates, 2)))
colnames(args) <- c("x1", "x2")
head(args)

#>    x1   x2
#> 1 cyl disp
#> 2 cyl   hp
#> 3 cyl drat
#> 4 cyl   wt
#> 5 cyl qsec
#> 6 cyl   vs

# For each row of `args`, we want to fit a model using the
# covariates in the row.
# We can formulate this task in terms of a `drake` plan.
plan <- map_plan(args, "my_model_fit")
plan

#> # A tibble: 45 x 2
#>    target                command                                    
#>    <chr>                 <chr>                                      
#>  1 my_model_fit_d8f8c343 "my_model_fit(x1 = \"cyl\", x2 = \"disp\")"
#>  2 my_model_fit_41b5eb56 "my_model_fit(x1 = \"cyl\", x2 = \"hp\")"  
#>  3 my_model_fit_11bff127 "my_model_fit(x1 = \"cyl\", x2 = \"drat\")"
#>  4 my_model_fit_0ab13a43 "my_model_fit(x1 = \"cyl\", x2 = \"wt\")"  
#>  5 my_model_fit_e86c89cb "my_model_fit(x1 = \"cyl\", x2 = \"qsec\")"
#>  6 my_model_fit_402688d8 "my_model_fit(x1 = \"cyl\", x2 = \"vs\")"  
#>  7 my_model_fit_622ee39d "my_model_fit(x1 = \"cyl\", x2 = \"am\")"  
#>  8 my_model_fit_0f1efb11 "my_model_fit(x1 = \"cyl\", x2 = \"gear\")"
#>  9 my_model_fit_2951b593 "my_model_fit(x1 = \"cyl\", x2 = \"carb\")"
#> 10 my_model_fit_eb3f790d "my_model_fit(x1 = \"disp\", x2 = \"hp\")" 
#> # ... with 35 more rows

# The default target names are ugly, so we can supply
# our own names. You can either set an "id" column in `args`
# or set the `id` argument of `map_plan()`.
args$id <- paste0("fit_", args$x1, "_", args$x2)
plan <- map_plan(args, "my_model_fit")
plan

#> # A tibble: 45 x 2
#>    target       command                                    
#>    <chr>        <chr>                                      
#>  1 fit_cyl_disp "my_model_fit(x1 = \"cyl\", x2 = \"disp\")"
#>  2 fit_cyl_hp   "my_model_fit(x1 = \"cyl\", x2 = \"hp\")"  
#>  3 fit_cyl_drat "my_model_fit(x1 = \"cyl\", x2 = \"drat\")"
#>  4 fit_cyl_wt   "my_model_fit(x1 = \"cyl\", x2 = \"wt\")"  
#>  5 fit_cyl_qsec "my_model_fit(x1 = \"cyl\", x2 = \"qsec\")"
#>  6 fit_cyl_vs   "my_model_fit(x1 = \"cyl\", x2 = \"vs\")"  
#>  7 fit_cyl_am   "my_model_fit(x1 = \"cyl\", x2 = \"am\")"  
#>  8 fit_cyl_gear "my_model_fit(x1 = \"cyl\", x2 = \"gear\")"
#>  9 fit_cyl_carb "my_model_fit(x1 = \"cyl\", x2 = \"carb\")"
#> 10 fit_disp_hp  "my_model_fit(x1 = \"disp\", x2 = \"hp\")" 
#> # ... with 35 more rows

# Let's actually fit our models.
# In your own work, you probably will not need to
# create or supply a cache
cache <- storr::storr_environment()
make(plan, cache = cache)

#> target fit_cyl_disp

#> target fit_cyl_hp

#> target fit_cyl_drat

#> target fit_cyl_wt

#> target fit_cyl_qsec

#> target fit_cyl_vs

#> target fit_cyl_am

#> target fit_cyl_gear

#> target fit_cyl_carb

#> target fit_disp_hp

#> target fit_disp_drat

#> target fit_disp_wt

#> target fit_disp_qsec

#> target fit_disp_vs

#> target fit_disp_am

#> target fit_disp_gear

#> target fit_disp_carb

#> target fit_hp_drat

#> target fit_hp_wt

#> target fit_hp_qsec

#> target fit_hp_vs

#> target fit_hp_am

#> target fit_hp_gear

#> target fit_hp_carb

#> target fit_drat_wt

#> target fit_drat_qsec

#> target fit_drat_vs

#> target fit_drat_am

#> target fit_drat_gear

#> target fit_drat_carb

#> target fit_wt_qsec

#> target fit_wt_vs

#> target fit_wt_am

#> target fit_wt_gear

#> target fit_wt_carb

#> target fit_qsec_vs

#> target fit_qsec_am

#> target fit_qsec_gear

#> target fit_qsec_carb

#> target fit_vs_am

#> target fit_vs_gear

#> target fit_vs_carb

#> target fit_am_gear

#> target fit_am_carb

#> target fit_gear_carb

# Let's inspect one of our models.
readd(fit_cyl_disp, cache = cache)

#> 
#> Call:
#> lm(formula = as.formula(paste("mpg ~", x1, "+", x2)), data = mtcars)
#> 
#> Coefficients:
#> (Intercept)          cyl         disp  
#>    34.66099     -1.58728     -0.02058  
#>