Defunct function

Staged parallelism is removed from drake, so this function is moot. drake uses a much better parallel algorithm now.

parallel_stages(config, from_scratch = FALSE)

Arguments

config
A configuration list output by make() or drake_config().

from_scratch
Logical, whether to assume that the next make() will run from scratch so that all targets are attempted.

Value

A data frame of information spelling out how targets are divided into parallelizable stages (according to the stage column).

Details

Made defunct on May 4, 2018.

Examples

# Do not use this function.