Read the pseudo-random number generator seed of the project.

When a project is created with make() or drake_config(), the project's pseudo-random number generator seed is cached. Then, unless the cache is destroyed, the seeds of all the targets will deterministically depend on this one central seed. That way, reproducibility is protected, even under randomness.

read_drake_seed(path = getwd(), search = TRUE, cache = NULL,
  verbose = 1L)

Arguments

path
Root directory of the drake project, or if search is TRUE, either the project root or a subdirectory of the project. Ignored if a cache is supplied.

search
Logical. If TRUE, search parent directories to find the nearest drake cache. Otherwise, look in the current working directory only. Ignored if a cache is supplied.

cache
drake cache. See new_cache(). If supplied, path and search are ignored.

verbose
Logical or numeric, control printing to the console.

0 or FALSE: print nothing.

1 or TRUE: print only targets to build.

2: plus checks and cache info.

3: plus missing imports.

4: plus all imports.

5: plus execution and total build times for targets.

6: plus notifications when targets are being stored.

Value

An integer vector.

Examples

cache <- storr::storr_environment() # Just for the examples.
my_plan <- drake_plan(
  target1 = sqrt(1234),
  target2 = sample.int(n = 12, size = 1) + target1
)
tmp <- sample.int(1) # Needed to get a .Random.seed, but not for drake.
digest::digest(.Random.seed) # Fingerprint of the current R session's seed.

#> [1] "efa692843822f13b7aa9c49a0d298b11"

make(my_plan, cache = cache) # Run the project, build the targets.

#> target target1

#> target target2

digest::digest(.Random.seed) # Your session's seed did not change.

#> [1] "efa692843822f13b7aa9c49a0d298b11"

# drake uses a hard-coded seed if you do not supply one.
read_drake_seed(cache = cache)

#> [1] 0

readd(target2, cache = cache) # Randomly-generated target data.

#> [1] 42.12834

clean(target2, cache = cache) # Oops, I removed the data!

#> 
#> Attaching package: ‘tidyselect’

#> The following object is masked from ‘package:testthat’:
#> 
#>     matches

tmp <- sample.int(1) # Maybe the R session's seed also changed.
make(my_plan, cache = cache) # Rebuild target2.

#> target target2

# Same as before:
read_drake_seed(cache = cache)

#> [1] 0

readd(target2, cache = cache)

#> [1] 42.12834

# You can also supply a seed.
# If your project already exists, it must agree with the project's
# preexisting seed (default: 0)
clean(target2, cache = cache)
make(my_plan, cache = cache, seed = 0)

#> target target2

read_drake_seed(cache = cache)

#> [1] 0

readd(target2, cache = cache)

#> [1] 42.12834

# If you want to supply a different seed than 0,
# you need to destroy the cache and start over first.
clean(destroy = TRUE, cache = cache)
cache <- storr::storr_environment() # Just for the examples.
make(my_plan, cache = cache, seed = 1234)

#> target target1

#> target target2

read_drake_seed(cache = cache)

#> [1] 1234

readd(target2, cache = cache)

#> [1] 38.12834