Render a visualization using the data frames generated by dataframes_graph().

This function is called inside vis_drake_graph(), which typical users call more often.

render_drake_graph(graph_dataframes, file = character(0),
  layout = "layout_with_sugiyama", direction = "LR", hover = TRUE,
  main = graph_dataframes$default_title, selfcontained = FALSE,
  navigationButtons = TRUE, ncol_legend = 1, ...)

Arguments

graph_dataframes
list of data frames generated by dataframes_graph(). There should be 3 data frames: nodes, edges, and legend_nodes.

file
Name of HTML file to save the graph. If NULL or character(0), no file is saved and the graph is rendered and displayed within R.

layout
name of an igraph layout to use, such as 'layout_with_sugiyama' or 'layout_as_tree'. Be careful with 'layout_as_tree': the graph is a directed acyclic graph, but not necessarily a tree.

direction
an argument to visNetwork::visHierarchicalLayout() indicating the direction of the graph. Options include 'LR', 'RL', 'DU', and 'UD'. At the time of writing this, the letters must be capitalized, but this may not always be the case ;) in the future.

hover
logical, whether to show the command that generated the target when you hover over a node with the mouse. For imports, the label does not change with hovering.

main
character string, title of the graph

selfcontained
logical, whether to save the file as a self-contained HTML file (with external resources base64 encoded) or a file with external resources placed in an adjacent directory. If TRUE, pandoc is required.

navigationButtons
logical, whether to add navigation buttons with visNetwork::visInteraction(navigationButtons = TRUE)

ncol_legend
number of columns in the legend nodes. To remove the legend entirely, set ncol_legend to NULL or 0.

...
arguments passed to visNetwork().

Value

A visNetwork graph.

Examples

# NOT RUN {
test_with_dir("Quarantine side effects.", {
load_basic_example() # Get the code with drake_example("basic").
# Instead of jumpting right to vis_drake_graph(), get the data frames
# of nodes, edges, and legend nodes.
config <- drake_config(my_plan) # Internal configuration list
graph <- dataframes_graph(config)
# You can pass the data frames right to render_drake_graph()
# (as in vis_drake_graph()) or you can create
# your own custom visNewtork graph.
render_drake_graph(graph, width = '100%') # Width is passed to visNetwork.
})
# }