Try to repair a drake cache that is prone to throwing storr-related errors.

Sometimes, storr caches may have dangling orphaned files that prevent you from loading or cleaning. This function tries to remove those files so you can use the cache normally again.

rescue_cache(targets = NULL, path = NULL, search = NULL,
  verbose = 1L, force = FALSE, cache = drake::drake_cache(path =
  path, verbose = verbose), jobs = 1, garbage_collection = FALSE)

Arguments

targets
Character vector, names of the targets to rescue. As with many other drake utility functions, the word target is defined generally in this case, encompassing imports as well as true targets. If targets is NULL, everything in the cache is rescued.

path
Character. Set path to the path of a storr::storr_rds() cache to retrieve a specific cache generated by storr::storr_rds() or drake::new_cache(). If the path argument is NULL, drake_cache() searches up through parent directories to find a folder called .drake/.

search
Deprecated.

verbose
Integer, control printing to the console/terminal.

0: print nothing.

1: print targets, retries, and failures.

2: also show a spinner when preprocessing tasks are underway.

force
Deprecated.

cache
A storr cache object.

jobs
Number of jobs for light parallelism (disabled on Windows).

garbage_collection
Logical, whether to do garbage collection as a final step. See drake_gc() and clean() for details.

Value

Nothing.

See also

Examples

# NOT RUN {
isolate_example("Quarantine side effects.", {
if (suppressWarnings(require("knitr"))) {
load_mtcars_example() # Get the code with drake_example("mtcars").
make(my_plan) # Run the project, build targets. This creates the cache.
# Remove dangling cache files that could cause errors.
rescue_cache(jobs = 2)
# Alternatively, just rescue targets 'small' and 'large'.
# Rescuing specific targets is usually faster.
rescue_cache(targets = c("small", "large"))
}
})
# }