Run a unit test in a way that quarantines the side effects from your workspace and file system.

Typical users of drake should not need this function. It is exported so it can be used to quarantine the side effects of the examples in the help files.

test_with_dir(desc, ...)

Arguments

desc
character, description of the test

...
code to test

Value

nothing

Examples

# NOT RUN {
test_with_dir(
  "Write a file to a temporary folder",
  writeLines("hello", "world.txt")
)
file.exists("world.txt") # FALSE
# }