Create the igraph dependency network of your project.

This function returns an igraph object representing how the targets in your workflow plan data frame depend on each other. (help(package = "igraph")). To plot the graph, call to plot.igraph() on your graph, or just use vis_drake_graph() from the start.

build_drake_graph(plan = read_drake_plan(),
  targets = drake::possible_targets(plan), envir = parent.frame(),
  verbose = 1, jobs = 1, sanitize_plan = TRUE)

Arguments

plan
workflow plan data frame, same as for function make().

targets
names of targets to build, same as for function make().

envir
environment to import from, same as for function make().

verbose
logical, whether to output messages to the console.

jobs
number of jobs to accelerate the construction of the dependency graph. A light mclapply()-based parallelism is used if your operating system is not Windows.

sanitize_plan
logical, whether to sanitize the workflow plan first.

Value

An igraph object representing the workflow plan dependency network.

See also

vis_drake_graph()

Examples

# NOT RUN {
test_with_dir("Quarantine side effects.", {
load_basic_example() # Get the code with drake_example("basic").
# Make the igraph network connecting all the targets and imports.
g <- build_drake_graph(my_plan)
class(g) # "igraph"
})
# }