Enumerate cached targets or check if a target is cached.

Read/load a cached item with readd() or loadd().

cached(..., list = character(0), no_imported_objects = FALSE,
  path = getwd(), search = TRUE, cache = NULL, verbose = 1L,
  namespace = NULL, jobs = 1)

Arguments

...
Objects to load from the cache, as names (unquoted) or character strings (quoted). Similar to ... in remove().

list
Character vector naming objects to be loaded from the cache. Similar to the list argument of remove().

no_imported_objects
Logical, applies only when no targets are specified and a list of cached targets is returned. If no_imported_objects is TRUE, then cached() shows built targets (with commands) plus imported files, ignoring imported objects. Otherwise, the full collection of all cached objects will be listed. Since all your functions and all their global variables are imported, the full list of imported objects could get really cumbersome.

path
Root directory of the drake project, or if search is TRUE, either the project root or a subdirectory of the project. Ignored if a cache is supplied.

search
Logical. If TRUE, search parent directories to find the nearest drake cache. Otherwise, look in the current working directory only. Ignored if a cache is supplied.

cache
drake cache. See new_cache(). If supplied, path and search are ignored.

verbose
Logical or numeric, control printing to the console.

0 or FALSE: print nothing.

1 or TRUE: print only targets to build.

2: plus checks and cache info.

3: plus missing imports.

4: plus all imports.

5: plus execution and total build times for targets.

6: plus notifications when targets are being stored.

namespace
Character scalar, name of the storr namespace to use for listing objects.

jobs
Number of jobs/workers for parallel processing.

Value

Either a named logical indicating whether the given targets or cached or a character vector listing all cached items, depending on whether any targets are specified.

See also

Examples

# NOT RUN {
test_with_dir("Quarantine side effects.", {
load_mtcars_example() # Load drake's canonical example.
make(my_plan) # Run the project, build all the targets.
cached(list = 'reg1') # Is 'reg1' in the cache?
# List all the targets and imported files in the cache.
# Exclude R objects imported from your workspace.
cached(no_imported_objects = TRUE)
# List all targets and imports in the cache.
cached()
# Clean the main data.
clean()
# The targets and imports are gone.
cached()
# But there is still metadata.
build_times()
cached(namespace = "build_times")
# Clear that too.
clean(purge = TRUE)
cached(namespace = "build_times")
build_times()
})
# }