Find out why a target is out of date.

The dependency profile can give you a hint as to why a target is out of date. It can tell you if

at least one input file changed,

at least one output file changed,

or a non-file dependency changed. For this last part, the imports need to be up to date in the cache, which you can do with outdated() or make(skip_targets = TRUE). Unfortunately, deps_profile() does not currently get more specific than that.

deps_profile(target, config, character_only = FALSE)

Arguments

target
Name of the target.

config
Configuration list output by drake_config() or make().

character_only
Logical, whether to assume target is a character string rather than a symbol.

Value

A data frame of the old hashes and new hashes of the data frame, along with an indication of which hashes changed since the last make().

See also

Examples

# NOT RUN {
isolate_example("Quarantine side effects.", {
if (suppressWarnings(require("knitr"))) {
load_mtcars_example() # Load drake's canonical example.
make(my_plan) # Run the project, build the targets.
config <- drake_config(my_plan)
# Get some example dependency profiles of targets.
deps_profile(small, config = config)
# Change a dependency.
simulate <- function(x) {}
# Update the in-memory imports in the cache
# so deps_profile can detect changes to them.
# Changes to targets are already cached.
make(my_plan, skip_targets = TRUE)
# The dependency hash changed.
deps_profile(small, config = config)
}
})
# }