History and provenance

See the history and provenance of your targets: what you ran, when you ran it, the function arguments you used, and how to get old data back.

drake_history(cache = NULL, history = NULL, analyze = FALSE,
  make_imports = TRUE, do_prework = TRUE, verbose = TRUE)

Arguments

cache
An optional drake/storr cache. If not supplied, drake will try to find the default cache: a folder called .drake/ in the current working directory or one of its ancestor directories.

history
An optional txtq produced by make(history = TRUE). If not supplied, drake will try to find it next to the cache in a folder called .drake_history/.

analyze
Logical, whether to analyze drake_plan() commands for arguments to function calls. Could be slow because this requires parsing and analyzing lots of R code.

make_imports
Logical, whether to make the imports first. Set to FALSE to save some time and risk obsolete output.

do_prework
Whether to do the prework normally supplied to make().

verbose
Integer, control printing to the console/terminal.

0: print nothing.

1: print targets, retries, and failures.

2: also show a spinner when preprocessing tasks are underway.

Value

A data frame of target history.

Details

If analyze is TRUE, drake scans your drake_plan() commands for function arguments and mentions them in the history. A function argument shows up if and only if 1. It has length 1. 2. It is atomic, i.e. a base type: logical, integer, real, complex, character, or raw. 3. It is explicitly named in the function call, For example, x is detected as 1 in fn(list(x = 1)) but not f(list(1)). The exceptions are file_out(), file_in(), and knitr_in(). For example, filename is detected as "my_file.csv" in process_data(filename = file_in("my_file.csv")). NB: in process_data(filename = file_in("a", "b")) filename is not detected because the value must be atomic.

Examples

# NOT RUN {
isolate_example({
# First, let's iterate on a drake workflow.
load_mtcars_example()
make(my_plan, history = TRUE, verbose = 0L)
# Naturally, we'll make updates to our targets along the way.
reg2 <- function(d) {
  d$x2 <- d$x ^ 3
  lm(y ~ x2, data = d)
}
Sys.sleep(0.01)
make(my_plan, history = TRUE, verbose = 0L)
# The history is a data frame about all the recorded runs of your targets.
out <- drake_history(analyze = TRUE)
print(out)
# Let's use the history to recover the oldest version
# of our regression2_small target.
oldest_reg2_small <- max(which(out$target == "regression2_small"))
hash_oldest_reg2_small <- out[oldest_reg2_small, ]$hash
cache <- drake_cache()
cache$get_value(hash_oldest_reg2_small)
# If you run clean(), drake can still find all the targets.
clean(small)
drake_history()
# But if you run clean() with garbage collection,
# older versions of your targets may be gone.
clean(large, garbage_collection = TRUE)
drake_history()
})
# }