Create a workflow plan data frame for the plan argument of make().

Turns a named collection of target/command pairs into a workflow plan data frame for make() and check(). You can give the commands as named expressions, or you can use the list argument to supply them as character strings.

drake_plan(..., list = character(0), file_targets = FALSE,
  strings_in_dots = c("filenames", "literals"), tidy_evaluation = TRUE)

Arguments

...
A collection of symbols/targets with commands assigned to them. See the examples for details.

list
A named list of targets, where the values are commands.

file_targets
logical, whether the targets should be (single-quoted) external file targets.

strings_in_dots
Character scalar, how to treat quoted character strings in the commands specified through .... Set to "filenames" to treat all these strings as external file targets/imports (single-quoted), or to "literals" to treat them all as literal strings (double-quoted). Unfortunately, because of how R deparses code, you cannot simply leave literal quotes alone in the ... argument. R will either convert all these quotes to single quotes or double quotes. Literal quotes in the list argument are left alone.

tidy_evaluation
logical, whether to use tidy evaluation such as quasiquotation when evaluating commands passed through the free-form ... argument.

Value

A data frame of targets and commands.

Details

A workflow plan data frame is a data frame with a target column and a command column. Targets are the objects and files that drake generates, and commands are the pieces of R code that produce them.

For file inputs and targets, drake uses single quotes. Double quotes are reserved for ordinary strings. The distinction is important because drake thinks about how files, objects, targets, etc. depend on each other. Quotes in the list argument are left alone, but R messes with quotes when it parses the free-form arguments in ..., so use the strings_in_dots argument to control the quoting in ....

Examples

# Create example workflow plan data frames for make()
drake_plan(small = simulate(5), large = simulate(50))

#>   target      command
#> 1  small  simulate(5)
#> 2  large simulate(50)

# Commands can be multi-line code chunks.
small_plan <- drake_plan(
  small_target = {
    local_object <- 1 + 1
    2 + sqrt(local_object)
  }
)
small_plan

#>         target                                                     command
#> 1 small_target {\n    local_object <- 1 + 1\n    2 + sqrt(local_object)\n}

cache <- storr::storr_environment() # Avoid writing files in examples.
make(small_plan, cache = cache) # Most users don't need the cache argument.

#> connect 2 imports: small_plan, cache

#> connect 1 target: small_target

#> Loading required package: methods

#> check 1 item: sqrt

#> check 1 item: small_target

#> target small_target

cached(cache = cache)

#> [1] "small_target" "sqrt"        

readd(small_target, cache = cache)

#> [1] 3.414214

# local_object only applies to the code chunk.
ls() # your environment is protected (local_object not found)

#> [1] "cache"        "small_plan"   "small_target"

rm(small_target)
# For tighter control over commands, use the `list` argument.
# Single quotes are file names, double quotes are oridinary strings.
# To actually run the little example workflow below,
# you need a file called `my_file.xlsx``
# with an Excel sheet called `second_sheet`.
# You also need to install the xlsx package.
drake_plan(
  list = c(
    dataset = "readxl::read_excel(path = 'my_file.xlsx', sheet = \"second_sheet\")" # nolint
  )
)

#>    target                                                           command
#> 1 dataset readxl::read_excel(path = 'my_file.xlsx', sheet = "second_sheet")

# Output targets can also be files,
# but the target names must have single quotes around them.
mtcars_plan <- drake_plan(
  output_file.csv = write.csv(mtcars, "output_file.csv"),
  file_targets = TRUE,
  strings_in_dots = "literals"
)
mtcars_plan

#>              target                              command
#> 1 'output_file.csv' write.csv(mtcars, "output_file.csv")

# make(mtcars_plan) # Would write output_file.csv. # nolint
# In the free-form `...` argument
# drake_plan() uses tidy evaluation to figure out your commands.
# For example, it respects the quasiquotation operator `!!`
# when it figures out what your code should be.
# Suppress this with `tidy_evaluation = FALSE` or
# with the `list` argument.
my_variable <- 5
drake_plan(
  a = !!my_variable,
  b = !!my_variable + 1,
  list = c(d = "!!my_variable")
)

#>   target       command
#> 1      a             5
#> 2      b             6
#> 3      d !!my_variable

drake_plan(
  a = !!my_variable,
  b = !!my_variable + 1,
  list = c(d = "!!my_variable"),
  tidy_evaluation = FALSE
)

#>   target             command
#> 1      a     !(!my_variable)
#> 2      b !(!my_variable + 1)
#> 3      d       !!my_variable

# For instances of !! that remain unevaluated in the workflow plan,
# make() will run these commands in tidy fashion,
# evaluating the !! operator using the environment you provided.