Use wildcard templating to create a workflow plan data frame from a template data frame.

Wildcards are no longer recommended. Consider using transformations instead. Visit https://ropenscilabs.github.io/drake-manual/plans.html#large-plans for the details.

evaluate_plan(plan, rules = NULL, wildcard = NULL, values = NULL,
  expand = TRUE, rename = expand, trace = FALSE,
  columns = "command", sep = "_")

Arguments

plan
Workflow plan data frame, similar to one produced by drake_plan().

rules
Named list with wildcards as names and vectors of replacements as values. This is a way to evaluate multiple wildcards at once. When not NULL, rules overrules wildcard and values if not NULL.

wildcard
Character scalar denoting a wildcard placeholder.

values
Vector of values to replace the wildcard in the drake instructions. Will be treated as a character vector. Must be the same length as plan$command if expand is TRUE.

expand
If TRUE, create a new rows in the workflow plan data frame if multiple values are assigned to a single wildcard. If FALSE, each occurrence of the wildcard is replaced with the next entry in the values vector, and the values are recycled.

rename
Logical, whether to rename the targets based on the values supplied for the wildcards (based on values or rules).

trace
Logical, whether to add columns that trace the wildcard expansion process. These new columns indicate which targets were evaluated and with which wildcards.

columns
Character vector of names of columns to look for and evaluate the wildcards.

sep
Character scalar, separator for the names of the new targets generated. For example, in evaluate_plan(drake_plan(x = sqrt(y__)), list(y__ = 1:2), sep = "."), the names of the new targets are x.1 and x.2.

Value

A workflow plan data frame with the wildcards evaluated.

Details

The commands in workflow plan data frames can have wildcard symbols that can stand for datasets, parameters, function arguments, etc. These wildcards can be evaluated over a set of possible values using evaluate_plan().

Specify a single wildcard with the wildcard and values arguments. In each command, the text in wildcard will be replaced by each value in values in turn. Specify multiple wildcards with the rules argument, which overrules wildcard and values if not NULL. Here, rules should be a list with wildcards as names and vectors of possible values as list elements.

See also

Examples

# Create the part of the workflow plan for the datasets.
datasets <- drake_plan(
  small = simulate(5),
  large = simulate(50)
)
# Create a template workflow plan for the analyses.
methods <- drake_plan(
  regression1 = reg1(dataset__),
  regression2 = reg2(dataset__)
)
# Evaluate the wildcards in the template
# to produce the actual part of the workflow plan
# that encodes the analyses of the datasets.
# Create one analysis for each combination of dataset and method.
evaluate_plan(
  methods,
  wildcard = "dataset__",
  values = datasets$target
)

#> The interface at https://ropenscilabs.github.io/drake-manual/plans.html#large-plans is better than evaluate_plan(), map_plan(), gather_by(), etc.

#> # A tibble: 4 x 2
#>   target            command    
#>   <chr>             <chr>      
#> 1 regression1_small reg1(small)
#> 2 regression1_large reg1(large)
#> 3 regression2_small reg2(small)
#> 4 regression2_large reg2(large)

# Only choose some combinations of dataset and analysis method.
ans <- evaluate_plan(
  methods,
  wildcard = "dataset__",
  values = datasets$target,
  expand = FALSE
)
ans

#> # A tibble: 2 x 2
#>   target      command    
#>   <chr>       <chr>      
#> 1 regression1 reg1(small)
#> 2 regression2 reg2(large)

# For the complete workflow plan, row bind the pieces together.
my_plan <- rbind(datasets, ans)
my_plan

#> # A tibble: 4 x 2
#>   target      command     
#>   <chr>       <chr>       
#> 1 small       simulate(5) 
#> 2 large       simulate(50)
#> 3 regression1 reg1(small) 
#> 4 regression2 reg2(large) 

# Wildcards for evaluate_plan() do not need the double-underscore suffix.
# Any valid symbol will do.
plan <- drake_plan(
  numbers = sample.int(n = `{Number}`, size = ..size)
)
evaluate_plan(
  plan,
  rules = list(
    "`{Number}`" = c(10, 13),
    ..size = c(3, 4)
  )
)

#> # A tibble: 4 x 2
#>   target       command                     
#>   <chr>        <chr>                       
#> 1 numbers_10_3 sample.int(n = 10, size = 3)
#> 2 numbers_10_4 sample.int(n = 10, size = 4)
#> 3 numbers_13_3 sample.int(n = 13, size = 3)
#> 4 numbers_13_4 sample.int(n = 13, size = 4)

# Workflow plans can have multiple wildcards.
# Each combination of wildcard values will be used
# Except when expand is FALSE.
x <- drake_plan(draws = sample.int(n = N, size = Size))
evaluate_plan(x, rules = list(N = 10:13, Size = 1:2))

#> # A tibble: 8 x 2
#>   target     command                     
#>   <chr>      <chr>                       
#> 1 draws_10_1 sample.int(n = 10, size = 1)
#> 2 draws_10_2 sample.int(n = 10, size = 2)
#> 3 draws_11_1 sample.int(n = 11, size = 1)
#> 4 draws_11_2 sample.int(n = 11, size = 2)
#> 5 draws_12_1 sample.int(n = 12, size = 1)
#> 6 draws_12_2 sample.int(n = 12, size = 2)
#> 7 draws_13_1 sample.int(n = 13, size = 1)
#> 8 draws_13_2 sample.int(n = 13, size = 2)

# You can use wildcards on columns other than "command"
evaluate_plan(
  drake_plan(
    x = target("1 + 1", cpu = "any"),
    y = target("sqrt(4)", cpu = "always"),
    z = target("sqrt(16)", cpu = "any")
  ),
  rules = list(always = 1:2),
  columns = c("command", "cpu")
)

#> # A tibble: 4 x 3
#>   target command        cpu  
#>   <chr>  <chr>          <chr>
#> 1 x      "\"1 + 1\""    any  
#> 2 y_1    "\"sqrt(4)\""  1    
#> 3 y_2    "\"sqrt(4)\""  2    
#> 4 z      "\"sqrt(16)\"" any  

# With the `trace` argument,
# you can generate columns that show how the wildcards
# were evaluated.
plan <- drake_plan(x = sample.int(n__), y = sqrt(n__))
plan <- evaluate_plan(plan, wildcard = "n__", values = 1:2, trace = TRUE)
print(plan)

#> # A tibble: 4 x 4
#>   target command       n__   n___from
#>   <chr>  <chr>         <chr> <chr>   
#> 1 x_1    sample.int(1) 1     x       
#> 2 x_2    sample.int(2) 2     x       
#> 3 y_1    sqrt(1)       1     y       
#> 4 y_2    sqrt(2)       2     y       

# With the `trace` argument,
# you can generate columns that show how the wildcards
# were evaluated. Then you can visualize the wildcard groups
# as clusters.
plan <- drake_plan(x = sqrt(n__), y = sample.int(n__))
plan <- evaluate_plan(plan, wildcard = "n__", values = 1:2, trace = TRUE)
print(plan)

#> # A tibble: 4 x 4
#>   target command       n__   n___from
#>   <chr>  <chr>         <chr> <chr>   
#> 1 x_1    sqrt(1)       1     x       
#> 2 x_2    sqrt(2)       2     x       
#> 3 y_1    sample.int(1) 1     y       
#> 4 y_2    sample.int(2) 2     y       

cache <- storr::storr_environment()
config <- drake_config(plan, cache = cache)
# NOT RUN {
if (requireNamespace("visNetwork", quietly = TRUE)) {
vis_drake_graph(config, group = "n__", clusters = "1")
vis_drake_graph(config, group = "n__", clusters = c("1", "2"))
make(plan, targets = c("x_1", "y_2"), cache = cache)
# Optionally cluster on columns supplied by `drake_graph_info()$nodes`.
vis_drake_graph(config, group = "status", clusters = "up to date")
}
# }