Gather multiple groupings of targets

Perform several calls to gather_plan() based on groupings from columns in the plan, and then row-bind the new targets to the plan.

gather_by(plan, ..., prefix = "target", gather = "list")

Arguments

plan
workflow plan data frame of prespecified targets

...
Symbols, columns of plan to define target groupings passed to dplyr::group_by(). A gather_plan() call is applied for each grouping. Groupings with all NAs in the selector variables are ignored.

prefix
character, prefix for naming the new targets. Suffixes are generated from the values of the columns specified in ....

gather
function used to gather the targets. Should be one of list(...), c(...), rbind(...), or similar.

See also

Examples

plan <- drake_plan(
  data = get_data(),
  informal_look = inspect_data(data, mu = mu__),
  bayes_model = bayesian_model_fit(data, prior_mu = mu__)
)
plan <- evaluate_plan(plan, rules = list(mu__ = 1:2), trace = TRUE)
gather_by(plan, mu___from, gather = "dplyr::bind_rows")

#> # A tibble: 7 x 4
#>   target         command                                      mu__  mu___from  
#>   <chr>          <chr>                                        <chr> <chr>      
#> 1 data           get_data()                                   <NA>  <NA>       
#> 2 informal_look… inspect_data(data, mu = 1)                   1     informal_l…
#> 3 informal_look… inspect_data(data, mu = 2)                   2     informal_l…
#> 4 bayes_model_1  bayesian_model_fit(data, prior_mu = 1)       1     bayes_model
#> 5 bayes_model_2  bayesian_model_fit(data, prior_mu = 2)       2     bayes_model
#> 6 target_bayes_… dplyr::bind_rows(bayes_model_1 = bayes_mode… <NA>  bayes_model
#> 7 target_inform… dplyr::bind_rows(informal_look_1 = informal… <NA>  informal_l…

gather_by(plan, mu__, mu___from, prefix = "x")

#> # A tibble: 9 x 4
#>   target            command                                 mu__  mu___from    
#>   <chr>             <chr>                                   <chr> <chr>        
#> 1 data              get_data()                              <NA>  <NA>         
#> 2 informal_look_1   inspect_data(data, mu = 1)              1     informal_look
#> 3 informal_look_2   inspect_data(data, mu = 2)              2     informal_look
#> 4 bayes_model_1     bayesian_model_fit(data, prior_mu = 1)  1     bayes_model  
#> 5 bayes_model_2     bayesian_model_fit(data, prior_mu = 2)  2     bayes_model  
#> 6 x_1_bayes_model   list(bayes_model_1 = bayes_model_1)     1     bayes_model  
#> 7 x_1_informal_look list(informal_look_1 = informal_look_1) 1     informal_look
#> 8 x_2_bayes_model   list(bayes_model_2 = bayes_model_2)     2     bayes_model  
#> 9 x_2_informal_look list(informal_look_2 = informal_look_2) 2     informal_look

reduce_by(plan, mu___from, begin = "list(", end = ")", op = ", ")

#> # A tibble: 7 x 4
#>   target               command                               mu__  mu___from   
#>   <chr>                <chr>                                 <chr> <chr>       
#> 1 data                 get_data()                            <NA>  <NA>        
#> 2 informal_look_1      inspect_data(data, mu = 1)            1     informal_lo…
#> 3 informal_look_2      inspect_data(data, mu = 2)            2     informal_lo…
#> 4 bayes_model_1        bayesian_model_fit(data, prior_mu = … 1     bayes_model 
#> 5 bayes_model_2        bayesian_model_fit(data, prior_mu = … 2     bayes_model 
#> 6 target_bayes_model   list(bayes_model_1, bayes_model_2)    <NA>  bayes_model 
#> 7 target_informal_look list(informal_look_1, informal_look_… <NA>  informal_lo…

reduce_by(plan, mu__, mu___from)

#> # A tibble: 5 x 4
#>   target          command                                mu__  mu___from    
#>   <chr>           <chr>                                  <chr> <chr>        
#> 1 data            get_data()                             <NA>  <NA>         
#> 2 informal_look_1 inspect_data(data, mu = 1)             1     informal_look
#> 3 informal_look_2 inspect_data(data, mu = 2)             2     informal_look
#> 4 bayes_model_1   bayesian_model_fit(data, prior_mu = 1) 1     bayes_model  
#> 5 bayes_model_2   bayesian_model_fit(data, prior_mu = 2) 2     bayes_model