Gather multiple groupings of targets

Perform several calls to gather_plan() based on groupings from columns in the plan, and then row-bind the new targets to the plan.

gather_by(plan, ..., prefix = "target", gather = "list",
  append = TRUE, filter = NULL, sep = "_")

Arguments

plan
workflow plan data frame of prespecified targets

...
Symbols, columns of plan to define target groupings. A gather_plan() call is applied for each grouping. Groupings with all NAs in the selector variables are ignored.

prefix
character, prefix for naming the new targets. Suffixes are generated from the values of the columns specified in ....

gather
function used to gather the targets. Should be one of list(...), c(...), rbind(...), or similar.

append
logical. If TRUE, the output will include the original rows in the plan argument. If FALSE, the output will only include the new targets and commands.

filter
an expression like you would pass to dplyr::filter(). The rows for which filter evaluates to TRUE will be gathered, and the rest will be excluded from gathering. Why not just call dplyr::filter() before gather_by()? Because gather_by(append = TRUE, filter = my_column == "my_value") gathers on some targets while including all the original targets in the output. See the examples for a demonstration.

sep
character scalar, delimiter for creating the names of new targets

Value

A workflow plan data frame.

See also

Examples

plan <- drake_plan(
  data = get_data(),
  informal_look = inspect_data(data, mu = mu__),
  bayes_model = bayesian_model_fit(data, prior_mu = mu__)
)
plan <- evaluate_plan(plan, rules = list(mu__ = 1:2), trace = TRUE)
plan

#> # A tibble: 5 x 4
#>   target          command                                mu__  mu___from    
#>   <chr>           <chr>                                  <chr> <chr>        
#> 1 data            get_data()                             <NA>  <NA>         
#> 2 informal_look_1 inspect_data(data, mu = 1)             1     informal_look
#> 3 informal_look_2 inspect_data(data, mu = 2)             2     informal_look
#> 4 bayes_model_1   bayesian_model_fit(data, prior_mu = 1) 1     bayes_model  
#> 5 bayes_model_2   bayesian_model_fit(data, prior_mu = 2) 2     bayes_model  

gather_by(plan, mu___from, gather = "rbind")

#> # A tibble: 8 x 4
#>   target         command                                       mu__  mu___from  
#>   <chr>          <chr>                                         <chr> <chr>      
#> 1 data           get_data()                                    <NA>  <NA>       
#> 2 informal_look… inspect_data(data, mu = 1)                    1     informal_l…
#> 3 informal_look… inspect_data(data, mu = 2)                    2     informal_l…
#> 4 bayes_model_1  bayesian_model_fit(data, prior_mu = 1)        1     bayes_model
#> 5 bayes_model_2  bayesian_model_fit(data, prior_mu = 2)        2     bayes_model
#> 6 target_bayes_… rbind(bayes_model_1 = bayes_model_1, bayes_m… <NA>  bayes_model
#> 7 target_inform… rbind(informal_look_1 = informal_look_1, inf… <NA>  informal_l…
#> 8 target_NA      rbind(data = data)                            <NA>  <NA>       

gather_by(plan, mu___from, append = TRUE)

#> # A tibble: 8 x 4
#>   target         command                                       mu__  mu___from  
#>   <chr>          <chr>                                         <chr> <chr>      
#> 1 data           get_data()                                    <NA>  <NA>       
#> 2 informal_look… inspect_data(data, mu = 1)                    1     informal_l…
#> 3 informal_look… inspect_data(data, mu = 2)                    2     informal_l…
#> 4 bayes_model_1  bayesian_model_fit(data, prior_mu = 1)        1     bayes_model
#> 5 bayes_model_2  bayesian_model_fit(data, prior_mu = 2)        2     bayes_model
#> 6 target_bayes_… list(bayes_model_1 = bayes_model_1, bayes_mo… <NA>  bayes_model
#> 7 target_inform… list(informal_look_1 = informal_look_1, info… <NA>  informal_l…
#> 8 target_NA      list(data = data)                             <NA>  <NA>       

gather_by(plan, mu___from, append = FALSE)

#> # A tibble: 3 x 3
#>   target          command                                            mu___from  
#>   <chr>           <chr>                                              <chr>      
#> 1 target_bayes_m… list(bayes_model_1 = bayes_model_1, bayes_model_2… bayes_model
#> 2 target_informa… list(informal_look_1 = informal_look_1, informal_… informal_l…
#> 3 target_NA       list(data = data)                                  <NA>       

gather_by(plan, mu__, mu___from, prefix = "x")

#> # A tibble: 10 x 4
#>    target            command                                 mu__  mu___from    
#>    <chr>             <chr>                                   <chr> <chr>        
#>  1 data              get_data()                              <NA>  <NA>         
#>  2 informal_look_1   inspect_data(data, mu = 1)              1     informal_look
#>  3 informal_look_2   inspect_data(data, mu = 2)              2     informal_look
#>  4 bayes_model_1     bayesian_model_fit(data, prior_mu = 1)  1     bayes_model  
#>  5 bayes_model_2     bayesian_model_fit(data, prior_mu = 2)  2     bayes_model  
#>  6 x_1_bayes_model   list(bayes_model_1 = bayes_model_1)     1     bayes_model  
#>  7 x_2_bayes_model   list(bayes_model_2 = bayes_model_2)     2     bayes_model  
#>  8 x_1_informal_look list(informal_look_1 = informal_look_1) 1     informal_look
#>  9 x_2_informal_look list(informal_look_2 = informal_look_2) 2     informal_look
#> 10 x_NA_NA           list(data = data)                       <NA>  <NA>         

gather_by(plan) # Gather everything and append a row.

#> # A tibble: 6 x 4
#>   target       command                                         mu__  mu___from  
#>   <chr>        <chr>                                           <chr> <chr>      
#> 1 data         get_data()                                      <NA>  <NA>       
#> 2 informal_lo… inspect_data(data, mu = 1)                      1     informal_l…
#> 3 informal_lo… inspect_data(data, mu = 2)                      2     informal_l…
#> 4 bayes_model… bayesian_model_fit(data, prior_mu = 1)          1     bayes_model
#> 5 bayes_model… bayesian_model_fit(data, prior_mu = 2)          2     bayes_model
#> 6 target       list(data = data, informal_look_1 = informal_l… <NA>  <NA>       

# You can filter out the informal_look_* targets beforehand
# if you only want the bayes_model_* ones to be gathered.
# The advantage here is that if you also need `append = TRUE`,
# only the bayes_model_* targets will be gathered, but
# the informal_look_* targets will still be included
# in the output.
gather_by(
  plan,
  mu___from,
  append = TRUE,
  filter = mu___from == "bayes_model"
)

#> # A tibble: 6 x 4
#>   target         command                                       mu__  mu___from  
#>   <chr>          <chr>                                         <chr> <chr>      
#> 1 data           get_data()                                    <NA>  <NA>       
#> 2 informal_look… inspect_data(data, mu = 1)                    1     informal_l…
#> 3 informal_look… inspect_data(data, mu = 2)                    2     informal_l…
#> 4 bayes_model_1  bayesian_model_fit(data, prior_mu = 1)        1     bayes_model
#> 5 bayes_model_2  bayesian_model_fit(data, prior_mu = 2)        2     bayes_model
#> 6 target_bayes_… list(bayes_model_1 = bayes_model_1, bayes_mo… <NA>  bayes_model