Write commands to combine several targets into one or more overarching targets.

Creates a new workflow plan to aggregate existing targets in the supplied plan.

gather_plan(plan = NULL, target = "target", gather = "list",
  append = FALSE)

Arguments

plan
workflow plan data frame of prespecified targets

target
name of the new aggregated target

gather
function used to gather the targets. Should be one of list(...), c(...), rbind(...), or similar.

append
logical. If TRUE, the output will include the original rows in the plan argument. If FALSE, the output will only include the new targets and commands.

Value

A workflow plan data frame that aggregates multiple prespecified targets into one additional target downstream.

See also

Examples

# Workflow plan for datasets:
datasets <- drake_plan(
  small = simulate(5),
  large = simulate(50)
)
# Create a new target that brings the datasets together.
gather_plan(datasets, target = "my_datasets", append = FALSE)

#> # A tibble: 1 x 2
#>   target      command                           
#>   <chr>       <chr>                             
#> 1 my_datasets list(small = small, large = large)

# This time, the new target just appends the rows of 'small' and 'large'
# into a single matrix or data frame.
gathered <- gather_plan(
  datasets,
  target = "aggregated_data",
  gather = "rbind",
  append = FALSE
)
gathered

#> # A tibble: 1 x 2
#>   target          command                            
#>   <chr>           <chr>                              
#> 1 aggregated_data rbind(small = small, large = large)

# For the complete workflow plan, row bind the pieces together.
bind_plans(datasets, gathered)

#> # A tibble: 3 x 2
#>   target          command                            
#>   <chr>           <chr>                              
#> 1 small           simulate(5)                        
#> 2 large           simulate(50)                       
#> 3 aggregated_data rbind(small = small, large = large)

# Alternatively, you can set `append = TRUE` to incorporate
# the new targets automatically.
gather_plan(
  datasets,
  target = "aggregated_data",
  gather = "rbind",
  append = TRUE
)

#> # A tibble: 3 x 2
#>   target          command                            
#>   <chr>           <chr>                              
#> 1 small           simulate(5)                        
#> 2 large           simulate(50)                       
#> 3 aggregated_data rbind(small = small, large = large)