Create a plan that maps a function to a grid of arguments.

map_plan() is like pmap_df() from the purrr package. It takes a function name and a grid of arguments, and it writes out all the commands calls to apply the function to each row of arguments.

map_plan(args, fun, id = "id", character_only = FALSE)

Arguments

args
a data frame (or better yet, a tibble) of function arguments to fun. Here, the column names should be the names of the arguments of fun, and each row of args corresponds to a call to fun.

fun
name of a function to apply the arguments row-by-row. Supply a symbol if character_only is FALSE and a character scalar otherwise.

id
name of an optional column in args giving the names of the targets. If not supplied, target names will be generated automatically. id should be a symbol if character_only is FALSE and a character scalar otherwise.

character_only
logical, whether to interpret the fun and id arguments as character scalars or symbols.

Value

A workflow plan data frame

See also

Examples

# For the full tutorial, visit
# https://ropenscilabs.github.io/drake-manual/plans.html#map_plan.
my_model_fit <- function(x1, x2, data){
  lm(as.formula(paste("mpg ~", x1, "+", x1)), data = data)
}
covariates <- setdiff(colnames(mtcars), "mpg")
args <- tibble::as_tibble(t(combn(covariates, 2)))
colnames(args) <- c("x1", "x2")
args$data <- "mtcars"
args$data <- rlang::syms(args$data)
args$id <- paste0("fit_", args$x1, "_", args$x2)
args

#> # A tibble: 45 x 4
#>    x1    x2    data     id          
#>    <chr> <chr> <list>   <chr>       
#>  1 cyl   disp  <symbol> fit_cyl_disp
#>  2 cyl   hp    <symbol> fit_cyl_hp  
#>  3 cyl   drat  <symbol> fit_cyl_drat
#>  4 cyl   wt    <symbol> fit_cyl_wt  
#>  5 cyl   qsec  <symbol> fit_cyl_qsec
#>  6 cyl   vs    <symbol> fit_cyl_vs  
#>  7 cyl   am    <symbol> fit_cyl_am  
#>  8 cyl   gear  <symbol> fit_cyl_gear
#>  9 cyl   carb  <symbol> fit_cyl_carb
#> 10 disp  hp    <symbol> fit_disp_hp 
#> # ... with 35 more rows

plan <- map_plan(args, my_model_fit)
plan

#> # A tibble: 45 x 2
#>    target       command                                                   
#>    <chr>        <chr>                                                     
#>  1 fit_cyl_disp "my_model_fit(x1 = \"cyl\", x2 = \"disp\", data = mtcars)"
#>  2 fit_cyl_hp   "my_model_fit(x1 = \"cyl\", x2 = \"hp\", data = mtcars)"  
#>  3 fit_cyl_drat "my_model_fit(x1 = \"cyl\", x2 = \"drat\", data = mtcars)"
#>  4 fit_cyl_wt   "my_model_fit(x1 = \"cyl\", x2 = \"wt\", data = mtcars)"  
#>  5 fit_cyl_qsec "my_model_fit(x1 = \"cyl\", x2 = \"qsec\", data = mtcars)"
#>  6 fit_cyl_vs   "my_model_fit(x1 = \"cyl\", x2 = \"vs\", data = mtcars)"  
#>  7 fit_cyl_am   "my_model_fit(x1 = \"cyl\", x2 = \"am\", data = mtcars)"  
#>  8 fit_cyl_gear "my_model_fit(x1 = \"cyl\", x2 = \"gear\", data = mtcars)"
#>  9 fit_cyl_carb "my_model_fit(x1 = \"cyl\", x2 = \"carb\", data = mtcars)"
#> 10 fit_disp_hp  "my_model_fit(x1 = \"disp\", x2 = \"hp\", data = mtcars)" 
#> # ... with 35 more rows

cache <- storr::storr_environment()
make(plan, verbose = FALSE, cache = cache)
readd(fit_cyl_disp, cache = cache)

#> 
#> Call:
#> lm(formula = as.formula(paste("mpg ~", x1, "+", x1)), data = data)
#> 
#> Coefficients:
#> (Intercept)          cyl  
#>      37.885       -2.876  
#>