Deprecated: show a ggraph/ggplot2 representation of your drake project.

Use drake_ggraph() instead.

static_drake_graph(config, build_times = "build", digits = 3,
  targets_only = FALSE, main = NULL, from = NULL, mode = c("out",
  "in", "all"), order = NULL, subset = NULL, make_imports = TRUE,
  from_scratch = FALSE, full_legend = FALSE, group = NULL,
  clusters = NULL)

Arguments

config
A drake_config() configuration list. You can get one as a return value from make() as well.

build_times
Character string or logical. If character, the choices are 1. "build": runtime of the command plus the time it take to store the target or import. 2. "command": just the runtime of the command. 3. "none": no build times. If logical, build_times selects whether to show the times from `build_times(..., type = "build")`` or use no build times at all. See build_times() for details.

digits
Number of digits for rounding the build times

targets_only
Logical, whether to skip the imports and only include the targets in the workflow plan.

main
Character string, title of the graph.

from
Optional collection of target/import names. If from is nonempty, the graph will restrict itself to a neighborhood of from. Control the neighborhood with mode and order.

mode
Which direction to branch out in the graph to create a neighborhood around from. Use "in" to go upstream, "out" to go downstream, and "all" to go both ways and disregard edge direction altogether.

order
How far to branch out to create a neighborhood around from. Defaults to as far as possible. If a target is in the neighborhood, then so are all of its custom file_out() files if show_output_files is TRUE. That means the actual graph order may be slightly greater than you might expect, but this ensures consistency between show_output_files = TRUE and show_output_files = FALSE.

subset
Optional character vector. Subset of targets/imports to display in the graph. Applied after from, mode, and order. Be advised: edges are only kept for adjacent nodes in subset. If you do not select all the intermediate nodes, edges will drop from the graph.

make_imports
Logical, whether to make the imports first. Set to FALSE to increase speed and risk using obsolete information.

from_scratch
Logical, whether to assume all the targets will be made from scratch on the next make(). Makes all targets outdated, but keeps information about build progress in previous make()s.

full_legend
Logical. If TRUE, all the node types are printed in the legend. If FALSE, only the node types used are printed in the legend.

group
Optional character scalar, name of the column used to group nodes into columns. All the columns names of your config$plan are choices. The other choices (such as "status") are column names in the nodes . To group nodes into clusters in the graph, you must also supply the clusters argument.

clusters
Optional character vector of values to cluster on. These values must be elements of the column of the nodes data frame that you specify in the group argument to drake_graph_info().

Value

A ggplot2 object, which you can modify with more layers, show with plot(), or save as a file with ggsave().

Details

Deprecated on 2018-25-07.

Examples

# See drake_ggraph()