Define custom columns in a drake_plan().

Not a user-side function. Must be called inside drake_plan().

target(command = NULL, ...)

Arguments

command
The command to build the target.

...
Optional columns of the plan for a given target. See the Columns section of this help file for a selection of special columns that drake understands.

Value

A one-row workflow plan data frame with the named arguments as columns.

Columns

drake_plan() creates a special data frame. At minimum, that data frame must have columns target and command with the target names and the R code chunks to build them, respectively.

You can add custom columns yourself, either with target() (e.g. drake_plan(y = target(f(x), transform = map(c(1, 2)), format = "fst"))) or by appending columns post-hoc (e.g. plan$col <- vals).

Some of these custom columns are special. They are optional, but drake looks for them at various points in the workflow.

transform: a call to map(), split(), cross(), or combine() to create and manipulate large collections of targets. Details: (https://ropenscilabs.github.io/drake-manual/plans.html#large-plans). # nolint

format: set a storage format to save big targets more efficiently. Most formats are faster than ordinary storage, and they consume far less memory. Available formats:

"fst": save big data frames fast. The target must be a data frame, and you must have the fst package installed.

"keras": save Keras models as HDF5 files. Requires the keras package.

"rds": save any object. This is similar to the default storage except we avoid creating a serialized copy of the entire target in memory. Requires R >= 3.5.0 so drake can use ALTREP.

trigger: rule to decide whether a target needs to run. It is recommended that you define this one with target(). Details: https://ropenscilabs.github.io/drake-manual/triggers.html.

hpc: logical values (TRUE/FALSE/NA) whether to send each target to parallel workers. Visit https://ropenscilabs.github.io/drake-manual/hpc.html#selectivity to learn more.

resources: target-specific lists of resources for a computing cluster. See https://ropenscilabs.github.io/drake-manual/hpc.html#advanced-options for details.

elapsed and cpu: number of seconds to wait for the target to build before timing out (elapsed for elapsed time and cpu for CPU time).

retries: number of times to retry building a target in the event of an error.

seed: an optional pseudo-random number generator (RNG) seed for each target. drake usually comes up with its own unique reproducible target-specific seeds using the global seed (the seed argument to make() and drake_config()) and the target names, but you can overwrite these automatic seeds. NA entries default back to drake's automatic seeds.

Keywords

drake_plan() understands special keyword functions for your commands. With the exception of target(), each one is a proper function with its own help file.

target(): give the target more than just a command. Using target(), you can apply a transformation (examples: https://ropenscilabs.github.io/drake-manual/plans.html#large-plans), # nolint supply a trigger (https://ropenscilabs.github.io/drake-manual/triggers.html), # nolint or set any number of custom columns.

map(): create multiple targets. Called inside target().

split(): create a target for each subset of data. Called inside target().

cross(): create a target for each combination of values. Called inside target().

combine(): aggregate groups of targets into other targets. Called inside target().

file_in(): declare an input file dependency.

file_out(): declare an output file to be produced when the target is built.

knitr_in(): declare a knitr file dependency such as an R Markdown (*.Rmd) or R LaTeX (*.Rnw) file.

ignore(): force drake to entirely ignore a piece of code: do not track it for changes and do not analyze it for dependencies.

no_deps(): tell drake to not track the dependencies of a piece of code. drake still tracks the code itself for changes.

drake_envir(): get the environment where drake builds targets. Intended for advanced custom memory management.

See also

Examples

# Use target() to create your own custom columns in a drake plan.
# See ?triggers for more on triggers.
drake_plan(
  website_data = target(
    download_data("www.your_url.com"),
    trigger = "always",
    custom_column = 5
  ),
  analysis = analyze(website_data)
)

#> # A tibble: 2 x 4
#>   target       command                           trigger       custom_column
#>   <chr>        <expr>                            <expr>                <dbl>
#> 1 website_data download_data("www.your_url.com") "always"                  5
#> 2 analysis     analyze(website_data)             NA_character_            NA

models <- c("glm", "hierarchical")
plan <- drake_plan(
  data = target(
    get_data(x),
    transform = map(x = c("simulated", "survey"))
  ),
  analysis = target(
    analyze_data(data, model),
    transform = cross(data, model = !!models, .id = c(x, model))
  ),
  summary = target(
    summarize_analysis(analysis),
    transform = map(analysis, .id = c(x, model))
  ),
  results = target(
    bind_rows(summary),
    transform = combine(summary, .by = data)
  )
)
plan

#> # A tibble: 12 x 2
#>    target                     command                                           
#>    <chr>                      <expr>                                            
#>  1 data_simulated             get_data("simulated")                            …
#>  2 data_survey                get_data("survey")                               …
#>  3 analysis_simulated_glm     analyze_data(data_simulated, "glm")              …
#>  4 analysis_simulated_hierar… analyze_data(data_simulated, "hierarchical")     …
#>  5 analysis_survey_glm        analyze_data(data_survey, "glm")                 …
#>  6 analysis_survey_hierarchi… analyze_data(data_survey, "hierarchical")        …
#>  7 summary_simulated_glm      summarize_analysis(analysis_simulated_glm)       …
#>  8 summary_simulated_hierarc… summarize_analysis(analysis_simulated_hierarchica…
#>  9 summary_survey_glm         summarize_analysis(analysis_survey_glm)          …
#> 10 summary_survey_hierarchic… summarize_analysis(analysis_survey_hierarchical) …
#> 11 results_data_simulated     bind_rows(summary_simulated_glm, summary_simulate…
#> 12 results_data_survey        bind_rows(summary_survey_glm, summary_survey_hier…

drake_plan_source(plan)

#> drake_plan(
#>   data_simulated = get_data("simulated"),
#>   data_survey = get_data("survey"),
#>   analysis_simulated_glm = analyze_data(data_simulated, "glm"),
#>   analysis_simulated_hierarchical = analyze_data(data_simulated, "hierarchical"),
#>   analysis_survey_glm = analyze_data(data_survey, "glm"),
#>   analysis_survey_hierarchical = analyze_data(data_survey, "hierarchical"),
#>   summary_simulated_glm = summarize_analysis(analysis_simulated_glm),
#>   summary_simulated_hierarchical = summarize_analysis(analysis_simulated_hierarchical),
#>   summary_survey_glm = summarize_analysis(analysis_survey_glm),
#>   summary_survey_hierarchical = summarize_analysis(analysis_survey_hierarchical),
#>   results_data_simulated = bind_rows(summary_simulated_glm, summary_simulated_hierarchical),
#>   results_data_survey = bind_rows(summary_survey_glm, summary_survey_hierarchical)
#> )