Define custom columns in a drake_plan().

Not a user-side function. Please use from within drake_plan() only.

target(command = NULL, ...)

Arguments

command
The command to build the target.

...
Named arguments specifying non-standard fields of the workflow plan.

Value

A one-row workflow plan data frame with the named arguments as columns.

See also

Examples

# Use target() to create your own custom columns in a drake plan.
# See ?triggers for more on triggers.
drake_plan(
  website_data = target(
    download_data("www.your_url.com"),
    trigger = "always",
    custom_column = 5
  ),
  analysis = analyze(website_data)
)

#> # A tibble: 2 x 4
#>   target       command                           trigger custom_column
#>   <chr>        <expr>                            <expr>          <dbl>
#> 1 website_data download_data("www.your_url.com") always              5
#> 2 analysis     analyze(website_data)             NA                 NA