Experimental: transform a plan.

This feature has a lot of promise, but it is still experimental. Please review your workflow with vis_drake_graph() before you run it.

transform_plan(plan, trace = FALSE)

Arguments

plan
Workflow plan data frame with a column for targets, a column for commands, and a column for transformations.

trace
Logical, whether to add columns to show what happened during target transformations, e.g. drake_plan(x = target(..., transform = ...), transform = TRUE).

Value

A transformed workflow plan data frame

Details

The transform_plan() function take an existing drake plan and applies the transformations in the optional "transform" column, expanding and gathering targets to create a larger plan. Usually this is done inside drake_plan(transform = TRUE), but transform_plan() on its own is useful if you generated multiple plans with drake_plan(transform = FALSE) and and want to combine and transform them later.

See also

Examples

plan1 <- drake_plan(
  analysis = target(
    analyze_data("source"),
    transform = cross(source = c(source1, source2))
  ),
  transform = FALSE
)
plan2 <- drake_plan(
  summaries = target(
    summarize_analyses(analysis),
    transform = reduce()
  ),
  transform = FALSE
)
plan <- bind_plans(plan1, plan2)
plan

#> # A tibble: 2 x 3
#>   target    command                      transform                          
#>   <chr>     <chr>                        <chr>                              
#> 1 analysis  "analyze_data(\"source\")"   cross(source = c(source1, source2))
#> 2 summaries summarize_analyses(analysis) reduce()                           

transform_plan(plan)

#> # A tibble: 3 x 2
#>   target         command                                                        
#>   <chr>          <chr>                                                          
#> 1 analysis_sour… "analyze_data(\"source\")"                                     
#> 2 analysis_sour… "analyze_data(\"source\")"                                     
#> 3 summaries      "summarize_analyses(list(analysis_source1 = analysis_source1, …

transform_plan(plan, trace = TRUE)

#> # A tibble: 3 x 5
#>   target      command                               source analysis    summaries
#>   <chr>       <chr>                                 <chr>  <chr>       <chr>    
#> 1 analysis_s… "analyze_data(\"source\")"            sourc… analysis_s… <NA>     
#> 2 analysis_s… "analyze_data(\"source\")"            sourc… analysis_s… <NA>     
#> 3 summaries   "summarize_analyses(list(analysis_so… <NA>   <NA>        summaries